% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_labels.R
\name{check_labels}
\alias{check_labels}
\title{Error check label files}
\usage{
check_labels(folder = NULL)
}
\arguments{
\item{folder}{= Path to a folder containing the labels to check. Can be NULL and use RStudio folder dialogue instead.}
}
\value{
A dataframe listing the label files checked and 1/0 indication of pass status for each check.
}
\description{
Given a folder of labelled audio files the function does some basic
checks that the label files meet required standards. Currently the function checks:
\itemize{
\item if the label start is numeric
\item if the label end is numeric
\item if all species codes are recognised
\item if all call type codes are recognised
\item if channels are present, if they are correctly specified
\item if there is a matching wav or mp3 file
}
}
