% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_labels.R
\name{collate_labels}
\alias{collate_labels}
\title{Scan a folder and collate any label file contents}
\usage{
collate_labels(folder = "E:/Data", include_subfolders = FALSE)
}
\arguments{
\item{folder}{= str, folder to be scanned}

\item{include_subfolders}{= bool, whether to scan recursively}
}
\value{
A dataframe with columns start, end, label, label_file name and original_audio_file name.
}
\description{
Scan a folder and collate any label file contents
}
\details{
Scan a folder to find Audacity label files and collate any labels.
}
