% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chunks_from_labels.R
\name{create_chunks_from_labels}
\alias{create_chunks_from_labels}
\title{Create audio chunks from long labelled audio}
\usage{
create_chunks_from_labels(
  path_in = NULL,
  path_out = NULL,
  primary_spp = NULL,
  chunk_duration = 5
)
}
\arguments{
\item{path_in}{= the folder where the audio+label files are located. If NULL, uses directory picker pop-up.}

\item{path_out}{= the folder where the extracted audio chunks and associated
labels will be saved. If NULL, uses directory picker pop-up.}

\item{primary_spp}{= the species code to use for the exported file names.
If NULL this will default to the first species listed in the label string}

\item{chunk_duration}{= desired length of the exported audio chunk}
}
\description{
This utility is for extracting fixed-length audio chunks from
longer soundscape files on the basis of instantaneous (zero-length) labels.
It assumes that labels are of the form \link{BT-S,GT-S,ST-F}.
}
