% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chunk.R
\name{extract_chunk}
\alias{extract_chunk}
\title{Extract audio chunk}
\usage{
extract_chunk(
  file_wav,
  file_chunk,
  start,
  end,
  chunk_duration,
  mono = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{file_wav}{= full path to the original audio file}

\item{file_chunk}{= full path and filename for the output chunk}

\item{start}{= start time of the detection in file_wav, in seconds}

\item{end}{= end time of the detection in file_wav, in seconds}

\item{chunk_duration}{= desired length of the chunk, in seconds. See Details}

\item{mono}{= should stereo recordings be exported as mono?}

\item{verbose}{= should messages be printed to console?}
}
\description{
Extract short audio chunk from a longer file. Useful for extracting classifier
detections from long passive audio files.
}
\details{
chunk_duration can be set to longer than the length of a detection. For example,
for a classifier that outputs detections of length 4s, setting chunk_duration to
7 will export a 7-second chunk centred on the 4s of interest, thereby giving
additional context to aid in verification, or for downstream training clips
with jitter. Note that if a chunk is at the start or end of a file, silence
is padded onto the start or end of the chunk to maintain chunk_duration.

Output directories will be created as required if not already present.
}
\examples{
\dontrun{
extract_chunk(file_wav = 'C:/long_audio_file.wav', file_chunk = 'C:/extracts/sparrow.wav', start = 4, end = 8, chunk_duration = 7)
}
}
\author{
Simon Gillings
}
