% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_duration.R
\name{get_duration}
\alias{get_duration}
\title{Get Audio File Duration}
\usage{
get_duration(wavfile)
}
\arguments{
\item{wavfile}{= the full path and filename of a wav file}
}
\value{
a list containing the file's sample rate, duration (seconds) and a
boolean flag for if the file is corrupt.
}
\description{
Get the duration (in seconds) of an audio file without reading the whole file.
}
\details{
Reading the whole of a large file simply to find our how long it is can be
inefficient. This function uses the tuneR function to read the header of
a wav file and uses the information therein to calculate the file's duration.
}
