% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_empty_labels.R
\name{make_empty_labels}
\alias{make_empty_labels}
\title{Create empty label files}
\usage{
make_empty_labels(overwrite = FALSE, subfolders = TRUE)
}
\arguments{
\item{overwrite}{= whether to overwrite existing label files (default FALSE)}

\item{subfolders}{= whether to make labels for audio in subfolders (default TRUE)}
}
\description{
For every audio file in a folder, create an empty text file of the same name
ready for populating with event labels.
}
\details{
This function is useful as a preparatory step when labelling audio files. For
a given folder of audio files the function will make an empty text file with
the same base name. Then, when labelling in Audacity it is quicker/safer to
save the labels to the existing empty file than to type in the file name in
manually. Optionally the user can specify to overwrite any existing label
files, and to scan recurisvely.
}
