% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels_from_filenames.R
\name{make_labels_from_filename}
\alias{make_labels_from_filename}
\title{Create fixed position label files using species in filename}
\usage{
make_labels_from_filename(
  position = NULL,
  calltype = NULL,
  overwrite = FALSE,
  subfolders = TRUE
)
}
\arguments{
\item{position}{= the position (in seconds) where the label should be created}

\item{calltype}{= the call type code to be assigned at position}

\item{overwrite}{= whether to overwrite existing label files (default FALSE)}

\item{subfolders}{= whether to make labels for audio in subfolders (default TRUE)}
}
\description{
For every audio file in a folder, create a label file of the same name with
the specified label position and inherit species code from the filename
}
\details{
For a given folder of audio files the function will
make an label file with the same base name, using the provided position
(seconds). It will pick the species code from the filename and add call type
provided. Optionally the user can specify to overwrite any existing label
files, and to scan recursively.
}
