% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_species_dict_for_python.R
\name{make_species_dict_for_python}
\alias{make_species_dict_for_python}
\title{Create species dictionary for Python workflows}
\usage{
make_species_dict_for_python(path_out)
}
\arguments{
\item{path_out, }{the location where the species dictionary should be written}
}
\description{
Takes the global_species_dictionary bundled in the BTOTools package
and exports all relevant taxa to make a lookup for Python workflows.
}
\examples{
make_species_dict_for_python(path_out = 'E://polyglotta/python/_packages/polyglotta')


}
