% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organise_labelled_clips.R
\name{organise_labelled_clips}
\alias{organise_labelled_clips}
\title{Organise labelled files into species folders}
\usage{
organise_labelled_clips(folder = NULL, path_library = NULL)
}
\arguments{
\item{folder}{= Path to a folder containing the audio and label files to be moved. Can be NULL and use RStudio folder dialogue instead.}

\item{path_library}{= Path to the top level of the audio library. Can be NULL and use RStudio folder dialogue instead.}
}
\description{
This objective of this function is to move audio files with paired label file
into appropriate 'species' folders within the audio library. It should be used
after running check_labels() to ensure compliance of labels with audio library
rules.
}
