% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_birdnet_results.R
\name{read_birdnet_csv}
\alias{read_birdnet_csv}
\title{Read BirdNET output}
\usage{
read_birdnet_csv(file, format = "Audacity", include_fname = FALSE)
}
\arguments{
\item{format}{= string, the format of the BirdNET output. One of Audacity,
Raven or csv (currently only Audacity supported)}

\item{include_fname}{= boolean, whether to include the label filename in the
output dataframe. This should be the same root as the audio filename so can
be used to track back to the audio to which the detections relate (but note
audio file extension is unknown).}
}
\value{
a dataframe containing detections and positions in the audio file
}
\description{
Reads and formats the standard output of a BirdNET run.
}
\details{
BirdNET outputs all detections in one of three formats: Audacity labels,
Raven selection tables or csv. For some of these formats the data are in an
unhelpful format (e.g. species English and scientific name are in a single
string). This function aims to simplify reading and standardising the format
of results read from BirdNET.
}
