% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_birdnet_results.R
\name{read_birdnet_results}
\alias{read_birdnet_results}
\title{Read and collate BirdNET outputs}
\usage{
read_birdnet_results(folder, subfolders = TRUE)
}
\arguments{
\item{folder}{= the full path to the folder for which results are to be collated/read}

\item{subfolders}{= bool, whether to include subfolders (default = TRUE)}
}
\value{
A dataframe of BirdNET detections with the following format. Note that
the last eight fields are only output by the R method:
\itemize{
\item original_wav = the full path to the origina audio file
\item start = start position in seconds of the detection
\item end = end position in seconds of the detection
\item birdnet_scientific_name = the scientific name of the species detected, following the taxonomy used by BirdNET
\item birdnet_english_name = the English name of the species detected, following the taxonomy used by BirdNET
\item score = BirdNET confidence score
\item lat = latitude (if supplied) for filtering
\item lon = longitude (if supplied) for filtering
\item week = week (if supplied) for filtering
\item overlap = overlap between chunks
\item sensitivity = sensitivity used in prediction
\item min_conf = minimum confidence filter applied
\item species_list = custom species list (if supplied)
\item model = classifier model version
}
}
\description{
This function can be used to read the different versions of BirdNET outputs and
compile them into a standard output format. It does not apply any filtering
beyond that already done in the BirdNET GUI.
}
\details{
BirdNET produces 5 format types, with some having the option to produce a
collated multi-wav output. The naming conventions for these are:
\itemize{
\item Audacity format: single file = WAVNAME.BirdNET.results.txt; No combined
output made.
\item R format: single = WAVNAME.BirdNET.results.r.csv; Combined format =
BirdNET_RTable.csv.
\item CSV format: single WAVNAME.BirdNET.results.csv; Combined format =
BirdNET_CombinedTable.csv.
}

The function assumes that ONLY ONE FORMAT is present in the passed folder,
although if a combined file is present, this is read preferentially to the individual files.
}
