% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_audiomoth_files.R
\name{rename_audiomoth_files}
\alias{rename_audiomoth_files}
\title{Rename AudioMoth audio files to audio library format}
\usage{
rename_audiomoth_files(
  folder = NULL,
  site,
  recordist = "BTO",
  include_subfolders = FALSE
)
}
\arguments{
\item{folder, }{string. Optional. Path to the audio files to be renamed. Can
be omitted and GUI pop-up used instead}

\item{site, }{string, a site name or lat-long string. No dashes permitted
as these are delimiters in the main file name format.}

\item{recordist, }{string, name of the recordist (or BTO if not provided).}

\item{include_subfolders, }{bool, whether audio files in subfolders should be
included, default FALSE}
}
\description{
Takes the standard format of an AudioMoth audio file and
converts to BTO audible recordings audio library format.
}
\details{
AudioMoth audio files typically are named YYYYMMDD_HHMMSS. In large projects
it is recommended to rename files to include location information to avoid
downstream confusion. This function will rename files in the format:
YYYYMMDD-HHMMSS-location-recordist-equipment.wav
where location and recordist are provided in the function call. The folder
of audio files to be renamed can be specified using a GUI interface or passed
as a parameter in the function call.
}
