% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_by_creation_date.R
\name{rename_by_creation_date}
\alias{rename_by_creation_date}
\title{Rename audio files on SD card using creation date}
\usage{
rename_by_creation_date(suffix, sep = "-")
}
\arguments{
\item{suffix, }{string, detailing a string to add to the date time prefix.
A typical use case will be a site name.}

\item{sep, }{string, the separator to use between datetime and suffix components.
Defaults to a hyphen. Underscore is another valid option.}
}
\description{
For audio files named consecutively, rename in a Pipeline-friendly format
using the recordings' creation datetime, which equates to the start datetime
of the recording.
}
\details{
This function will rename files in the format YYYYMMDD-HHMMSS-suffix.wav/mp3/flac,
where suffix is supplied in the function call. It uses the creation date of
the file. If the file is still on a memory card this will be the start time
of the recording. Beware if the file has been copied to a new location the
creation date no longer equates to the start time!
}
