% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_pipeline_files.R
\name{rename_pipeline_files}
\alias{rename_pipeline_files}
\title{Rename Pipeline audio files}
\usage{
rename_pipeline_files(
  folder = NULL,
  recordist = "BTOAP",
  species_code = NULL,
  include_subfolders = FALSE,
  include_label_files = TRUE
)
}
\arguments{
\item{folder, }{string. Optional. Path to the audio files to be renamed. Can
be omitted and GUI pop-up used instead}

\item{recordist, }{string, name of the recordist (BTOAP by default).}

\item{species_code, }{string, NULL or a a species code to be inserted in the filename}

\item{include_subfolders, }{bool, whether audio files in subfolders should be
included, default FALSE}

\item{include_label_file, }{bool, whether to rename label files in same folder}
}
\description{
Takes the standard format of filename produced by the Acoustic Pipeline and
converts to BTO audible recordings audio library format. Optionally can also
rename label (.txt) files for instances where labelling was done prior to
renaming of the audio file.
}
\details{
Pipeline files are of format: YYYYMMDD_HHMMSS-(latitude+longitude)-SUFFIX,
where SUFFIX is the name of the original file from which the extract was made.
This function will rename files in the format:
YYYYMMDD-HHMMSS-location-recordist-PREFIX.wav
where location is parsed from the lat-long and recordist is  provided in the
function call (but BTOAP by default). The folder of audio files to be renamed
can be specified using a GUI interface or passed as a parameter in the
function call.
}
