% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_xc_species_clips.R
\name{download_XC_species_clips}
\alias{download_XC_species_clips}
\title{Download Species Recordings from xeno-canto}
\usage{
download_XC_species_clips(path_out, code2ltr, n)
}
\arguments{
\item{path_out, }{string, the main path where files are to be saved.}

\item{code2ltr, }{string, the two letter code of the species required}

\item{n, }{numeric, the number of clips to return}
}
\description{
Uses the warbleR package and the xeno-canto API to download clips for a species.
}
\details{
Pass in a species name and a number of clips desired. Some filtering is currently
hard coded, e.g. only returns clips with quality A:B, removes nestlings and
birds not seen, and any clips with other species reported in the background.
It is possible these filters will result in zero returns.
}
