% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_birdnet_splist.R
\name{get_birdnet_splist}
\alias{get_birdnet_splist}
\title{Read BirdNET species list}
\usage{
get_birdnet_splist(path_birdnet = "F:/BirdNET-Analyzer", version = 2.4)
}
\arguments{
\item{path_birdnet, }{string, the path to where BirdNET Analyzer is installed,
e.g. "F:/BirdNET-Analyzer"}

\item{version, }{numeric, the BirdNET version to check, e.g. 2.4}
}
\description{
Get the full species list supported by the current version of BirdNET. Requires
a local installation of BirdNET.
}
